Better error handling in GlideExtensions
Fixes: https://github.com/jonathanpeppers/glidex/issues/5 (I hope)

Got a report of:

    Java.Lang.IllegalArgumentException: You cannot start a load for a destroyed activity occurred

I'm not able to reproduce the problem in the `glidex.forms.sample`,
however.

Doing some research, the Java community had a good recommendation:

https://github.com/bumptech/glide/issues/1484#issuecomment-365625087

I am not sure exactly *how* a destroyed, finishing, or non-Activity
could be encountered in `GlideExtensions`, but it does?

I also added a `try-catch` around the entire method, as there is
currently no other way a developer could handle this case themselves
if something is going wrong. I felt it better to log the problem, and
hopefully they file an issue.