IGlideHandler implementation
Fixes: https://github.com/jonathanpeppers/glidex/issues/28

This is my initial idea that allows developers to implement their own
calls to Glide in a simple way.

For example `RandomAlphaHandler` creates a `SimpleTarget` subclass
that sets a random alpha value on the `BitmapDrawable`.

When subclassing `SimpleTarget`, you can also override:

* `OnLoadStarted`
* `OnLoadFailed`

This should give developers flexibility for what happens if images
fail to load, etc.

TODO:
* Let someone else try out my API?
* Write a few other `IGlideHandler` examples