Cancel Glide requests in Dispose() (#53)
Fixes: https://github.com/jonathanpeppers/glidex/issues/45

Thanks @activa! I could finally reproduce this!

The problem here is we need to cancel any in-flight Glide requests
when the `ImageRenderer` is disposed. Otherwise, Java completes the
request against a C# object that is no longer alive.

There may be a future fix needed here for `ImageCell`, but I wasn't
able to reproduce a problem there.