ImageButton examples for VisualStateManager and Flexlayout (#74)
Context: https://github.com/jonathanpeppers/glidex/issues/71

You can see the issues by doing the following. 

Launch the app on device or in simulator and scroll down to the bottom
and select FlexLayout Image Buttons. This will display two image
buttons on the screen. When you tap on one you will notice it will
disappear for a second and then load in the pressed image. You can see
this better by holding the tap. After releasing the first time and
then tapping it again, it does not flash anymore however it completely
loses its positioning and starts resizing weird. If I tap only the top
one it shrinks down and up. If I tap the bottom one after the top one
it stretches to the side. Now if I tap the top one again it will
stretch to the side as well. On my personal app it takes about a whole
second for the pressed image to load but in this app its about .8
seconds but still noticeable. The point of this layout shows exactly
what I'm trying to do with my app that can have dynamic screen sizes
for ipad and iphone so it needs to be able to expand and shrink.

The second example is the carousel view image buttons. This one you
can see the image disappears and reappears on first use and then stays
loaded on additional uses. 

Co-authored-by: David Berroa <dberroa@lakesidellama.com>
Co-authored-by: Jonathan Peppers <jonathan.peppers@microsoft.com>