Fix flickering for VisualStateManager usage
Context: https://github.com/jonathanpeppers/glidex/issues/71

The `VisualStateManagerPage` displays an issue where you see an image
flicker for a moment when tapping a button for the first time.
Subsequent taps work fine.

I found that setting a `Placeholder` solved the problem completely,
and I don't see flickering any longer.

Not sure if this fully fixes #71 as there are some mentions of
`FlexLayout`.