Add IImageSourceHandler for NavigationPage.TitleIconImageSource
Fixes: https://github.com/jonathanpeppers/glidex/issues/69
Context: https://github.com/xamarin/Xamarin.Forms/issues/11676

There are certain Android APIs where `IImageViewHandler` is not
possible for Xamarin.Forms to be able to call. For example, if you
have an `IMenuItem`:

    menuItem.SetIcon (drawable);

`IImageViewHandler` requires an `ImageView` -- so Xamarin.Forms must
fall back to `IImageSourceHandler` for this to work at *all*.

Adding `IImageSourceHandler` required various refactoring, but things
seem to work as before. `IImageViewHandler` appears to be called by
Xamarin.Forms when it *should* be, and `IImageSourceHandler` is called
as a fallback.

Perhaps in MAUI, `IImageViewHandler` should have an additional method
for `IMenuItem`?

Co-authored-by: Yaser Moradi <ysmoradi@outlook.com>